+++ mini_app/common/src/config.rs
@@ -1,2 +1,2 @@
 pub const HTTP_SERVER_IP: &str = "127.0.0.1";
-pub const HTTP_SERVER_PORT: usize = 9001;
+pub const HTTP_SERVER_PORT: u16 = 9001;
