+++ src/config.rs
@@ -1,7 +1,7 @@
 use std::time;
 
 pub const ECHO_SERVER_IP: &str = "127.0.0.1";
-pub const ECHO_SERVER_PORT: usize = 8888;
+pub const ECHO_SERVER_PORT: u16 = 8888;
 
 pub const READ_BUF_SIZE: usize = 4;
 pub const MAX_WRITE_BUF_SIZE: usize = 16;
